# heka-identity-service/docs/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /proofs:
    get:
      operationId: ProofController_find
      parameters:
        - name: threadId
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Proof Records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProofRecordDto"
        "401":
          description: Unauthorized
      security: &a29
        - bearer: []
      summary: Get all proof records
      tags: &a30
        - Proofs
  /proofs/request:
    post:
      operationId: ProofController_request
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProofRequestDto"
      responses:
        "200":
          description: Proof Record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProofRecordDto"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable Entity
      security: *a29
      summary: Request a proof
      tags: *a30
  /proofs/{id}:
    get:
      operationId: ProofController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Proof Record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProofRecordDto"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security: *a29
      summary: Get a proof record
      tags: *a30
  /proofs/{id}/present:
    post:
      operationId: ProofController_present
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Proof Record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProofRecordDto"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "409":
          description: Conflict
        "422":
          description: Unprocessable Entity
      security: *a29
      summary: Present a proof
      tags: *a30
...
  /openid4vc/verifier:
    post:
      operationId: OpenId4VcVerifierController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenId4VcVerifierCreateDto"
      responses:
        "200":
          description: Verifier Record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenId4VcVerifierRecordDto"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable Entity
      security: &a35
        - bearer: []
      summary: Create a Verifier
      tags: &a36
        - OpenID4VC Verifier
    get:
      operationId: OpenId4VcVerifierController_find
      parameters:
        - name: publicVerifierId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Verifier records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenId4VcVerifierRecordDto"
        "401":
          description: Unauthorized
      security: *a35
      summary: Get verifier records
      tags: *a36
...
  /revocation-registries:
    post:
      operationId: RevocationRegistryController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRevocationRegistryRequest"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRevocationRegistryResponse"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security: &a42
        - bearer: []
      summary: Create revocation registry
      tags: &a43
        - Anoncreds Revocation Registries
    get:
      operationId: RevocationRegistryController_find
      parameters:
        - name: credDefId
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevocationRegistry"
        "401":
          description: Unauthorized
      security: *a42
      summary: Get created revocation registries
      tags: *a43
  /revocation-registries/{id}:
    get:
      operationId: RevocationRegistryController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: timestamp
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetRevocationRegistryResponse"
        "401":
          description: Unauthorized
      security: *a42
      summary: Get revocation registry
      tags: *a43
...
info:
  title: Heka Identity Service
  description: Heka Identity Service API
  version: 2.3.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ProofRevealedAttributeDto:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
      required:
        - name
        - value
    ProofRecordDto:
      type: object
      properties:
        id:
          type: string
        connectionId:
          type: string
        threadId:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        state:
          type: string
          enum:
            - proposal-sent
            - proposal-received
            - request-sent
            - request-received
            - presentation-sent
            - presentation-received
            - declined
            - abandoned
            - done
        isVerified:
          type: boolean
        revealedAttributes:
          type: array
          items:
            $ref: "#/components/schemas/ProofRevealedAttributeDto"
      required:
        - id
        - threadId
        - createdAt
        - state
    ProofRequestDto:
      type: object
      properties:
        connectionId:
          type: string
        comment:
          type: string
        requestNonRevokedProof:
          type: boolean
        request:
          oneOf:
            - $ref: "#/components/schemas/AnoncredsProofRequestDto"
            - $ref: "#/components/schemas/DifPresentationExchangeDto"
      required:
        - connectionId
        - request
...
    OpenId4VcVerifierCreateDto:
      type: object
      properties:
        publicVerifierId:
          type: string
      required:
        - publicVerifierId
    OpenId4VcVerifierRecordDto:
      type: object
      properties:
        id:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        type:
          type: string
        publicVerifierId:
          type: string
      required:
        - id
        - createdAt
        - type
        - publicVerifierId
...
    CreateRevocationRegistryRequest:
      type: object
      properties:
        credentialDefinitionId:
          type: string
        issuerId:
          type: string
        maximumCredentialNumber:
          type: number
      required:
        - credentialDefinitionId
        - issuerId
        - maximumCredentialNumber
    CreateRevocationRegistryResponse:
      type: object
      properties:
        revocationRegistryDefinitionId:
          type: string
        index:
          type: number
        maximumCredentialNumber:
          type: number
      required:
        - revocationRegistryDefinitionId
        - index
        - maximumCredentialNumber
    GetRevocationRegistryResponse:
      type: object
      properties:
        timestamp:
          type: number
        revocationStatusList:
          type: array
          items:
            type: string
      required:
        - timestamp
        - revocationStatusList
    RevocationRegistry:
      type: object
      properties:
        revocationRegistryDefinitionId:
          type: string
        index:
          type: number
        maximumCredentialNumber:
          type: number
      required:
        - revocationRegistryDefinitionId
        - index
        - maximumCredentialNumber
...
security:
  - bearer: []